fix compiler warning in linux

+++ Source/CNTKv2LibraryDll/Trainer.cpp
@@ -385,7 +385,7 @@ namespace CNTK
         else if (m_aggregatedLossFunction->Output().GetDataType() == DataType::Double)
             m_rootGradientValue->Data()->SetValue(1.0);
         else
-            m_rootGradientValue->Data()->SetValue(half(1.0));
+            m_rootGradientValue->Data()->SetValue(static_cast<half>(half(1.0)));
 
         for (const auto& parameter : m_learnerParameters)
             parameterGradients[parameter] = nullptr;
